% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_output.R
\name{createOutput}
\alias{createOutput}
\title{create_output}
\usage{
createOutput(output, path)
}
\arguments{
\item{output}{DataFrame}

\item{path}{The environment of project}
}
\value{

}
\description{
This function is called by the skeleton function and create the 
output doc in to Outpute File
}
