% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generarModelo.R
\name{entrenar_modelo}
\alias{entrenar_modelo}
\title{entrenar_modelo}
\usage{
entrenar_modelo(df_1, config)
}
\arguments{
\item{config}{}
}
\value{
output
}
\description{
esta función es llamada en GapminderApp y se encarga de partir datos y entrenar un modelo
}
\author{
leosanchezsoler
}
