% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GapminderApp.R
\name{GapminderApp}
\alias{GapminderApp}
\title{GapminderApp}
\usage{
GapminderApp(path)
}
\arguments{
\item{path}{}
}
\description{
esta función ejecuta el resto de funciones de la carpeta R
}
\author{
KapilDM
}
