% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leerDatos.R
\name{leer_datos}
\alias{leer_datos}
\title{leer_datos}
\usage{
leer_datos(config, path)
}
\arguments{
\item{path}{}
}
\value{
datos
}
\description{
función que lee archivos csv. Es llamada por la función main
}
\author{
leosanchezsoler
}
