% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leerDatos.R
\name{leer_target}
\alias{leer_target}
\title{leer_target}
\usage{
leer_target(config, path)
}
\arguments{
\item{path}{}
}
\value{
target
}
\description{
es llamada por main y se encarga de leer el target
}
\author{
leosanchezsoler
}
