% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_data.R
\name{reshape_data}
\alias{reshape_data}
\title{reshape_data}
\usage{
reshape_data(config, lista_pred_target)
}
\arguments{
\item{config}{List of configuration parameters}

\item{data}{list of datasets}
}
\value{

}
\description{
This function is called by skeleton and reshapes the df to a [n, 2] shape
}
